Broke context manager test out into it's own file.
